Make argument parsing in main() less error-prone.
Now, instead of being able to disable encryption by entering an empty password,
the -n/--no_password flag must be used. Encrypting with an empty password is
not allowed.

To create a new database, the -c flag must be passed. Otherwise, if the passed
file path doesn't exist, an error is printed. The -c flag won't overwrite an
existing file, either.