Bugfix in schema: replace INT with INTEGER.
Although SQLite recognizes INT as a column with INTEGER type affinity, the
primary key doesn't become an alias for the row id unless it is explicitly
declared INTEGER, and that is required to allow SQLite to assign row ids
automatically.

This doesn't affect any columns which aren't primary keys, but replacing all
occurrances of INT with INTEGER is less confusing.