Redefine meaning of nlink column in metadata table for directories.
We no longer count the '.' entry (even though we count it when filling in a
stat structure), so that unreachable directories will have nlink = 0.
Previously, unlinked directories would always have nlink >= 1, which makes it
impossible to find and purge them through the partial index on nlink = 0.
Since we don't store the self-link in the direntries table, it makes sense not
to count it in the nlink column either.