Don't CHECK() return value of sqlite3_reset() or sqlite3_clear_bindings().
Rationale: sqlite_reset() only returns SQLITE_OK if the last operation succeeded
so checking for this caused spurious crashes in cases where we handle errors by
setting the result to EIO before calling sqlite3_reset().

The result of sqlite3_clear_bindings() isn't clearly documented, and it doesn't
seem like it can fail in any way, so it's clearer not to check at all.