Created a separate test suite sqlfuse_external_tests.
This differs from sqlfuse_tests in that the former runs SQLFuse in a separate
thread, while the latter runs SQLFuse in a separate process. Running in a
separate process has the advantage that file locks will be released correctly
when the process exits, but running in a thread allows the test to manipulate
data structures used by SQLFuse (see e.g. test_blocksize()), so it's useful to
have both options.