Bug fix: when daemonizing, open sqlfs and intmap in the child process.
This is important because it is not safe to open a SQLite database in one
process, fork(), then use it in the child process. In particular, the child
process does not inherit all file locks from the parent, which means in our
case, it was possible to mount the same database in two locations at the same
time (even though the database was supposed to be opened in exclusive mode),
which would lead to database corruptions.

Also updated the CLI tests to test this case.