Remove warning printed when running in single-threaded mode.
Normally, fuse_main() would run in multi-threaded mode unless the -s option
is specified. However, we don't support multi-threading at all, and we don't
call fuse_main(), so this option does nothing.

We used to print a warning that the -s option was missing, but I think that is
more confusing than helpful, because it makes it seem like the mount has failed,
while in reality everything works perfectly fine.