Add unit tests for the sqlfs library itself.
This allows testing things on a lower level, without going through
libfuse. This also allows testing some edge cases that wouldn't occur in
practice, because FUSE handles them for us. It's still useful to test
these, because we want sqlfs to be self-contained, and not dependent on
FUSE for correctness.