Allow setting PRAGMA kdf_iter in tests.
This overrides the number of iterations of the key derivation function run,
which is useful to speed up the test when running with mcheck() leak checking.

This is basically the only use-case for this feature. kdf_iter should probably
not be overridden in any other case. In theory, we could make it an option for
sqlfuse, but then we might as well allow changing the hash function, too.