Add support for reading the password using an askpass program.
By default, the password is read from the terminal using libc's
getpass() function. This change adds the --askpass=command option, which
allows an arbitrary command to be executed to obtain the password. This
is particularly useful when invoking sqlfuse programatically from a GUI
frontend. The askpass command can be something like x11-ssh-askpass
which prompts for the password in an X window.