Support both SQLCipher v4 and v3 cipher parameters.
When the sqlcipher library is version 4 or higher, new databases will be
created with PRAGMA cipher_compatibility = 4, while existing databases will be
attempted to be opened as v4 first, and v3 second. This allows filesystem
databases created with v3 to be opened with v4, although it will be a little
slower.

If a v3 database file is opened with the v4 library, a warning will be printed.
Currently, there is no automatic way to upgrade the database, though the
instructions here can be used:
http://discuss.zetetic.net/t/upgrading-to-sqlcipher-4/3283/3